Moved the granule endpoints and cleaned up some modules.
