+++ search/package.json
@@ -29,7 +29,7 @@
   "scripts": {
     "start": "sls offline",
     "deploy": "sls deploy",
-    "lint": "eslint *.js ./lib/** ./tests/** --fix",
+    "lint": "eslint ./lib/** ./tests/** --fix",
     "test": "jest",
     "build:docs": "redoc-cli bundle \"docs/WFS3core+STAC.yaml\" -o docs/index.html --cdn --title \"STAC\"",
     "ci": "npm run lint && npm run test"
