--- search/tests/app.spec.js
@@ -1,7 +1,0 @@
-const app = require('../index.js');
-
-describe('Tests index', function () {
-  it('verifies successful response', async () => {
-    expect(app).toBeDefined();
-  });
-});
